Adding isAlignmentSupported() despite reservations
Towards https://github.com/browsermt/bergamot-translator/issues/77.

This function is added back to provide a state of the source during
collapse where nothing is expected to break at the extension, dropping
the UnifiedAPI from the loop. It is certain that extension cannot be
using it anywhere despite it's existence due to the functionality
missing.