WebAssembly: Disable wormhole and add CI via node
Removes WORMHOLE option from CMakeLists and corresponding builds on CI.

Add 'it works' CI for WebAssembly: import
https://gist.github.com/jelmervdl/a4c8b6b92ad88a885e1cbd51c6ad4902 and
attach it to CI.  NodeJS-14 is failing on trying to use the WebAssembly
binary. So we use node-16 independently setup.  This paves way for more
complicated testing for WebAssembly bindings in the future.